move the logger interface to under logging
